fix(giveback): auto-show funnel by saved causes, kill hero layout shift, split funnel
Onboarding:
- The funnel now shows automatically for any eligible visitor who hasn't
  saved causes (gated on selection.hasSavedCauses), not a local
  giveback:funnel_seen flag, so it keeps appearing until they pick.
- Hold the whole page body until the campaign status and saved picks
  resolve. The funnel is a full-screen overlay on the same background, so
  revealing the hero/tabs first flashed them before it covered them and
  popped the tabs in once the picks landed. Now only the shared background
  shows while resolving: no flash, no shift.

Hero layout shift:
- The mascot <img> had a fixed height but auto width and lazy loading, so
  it reflowed the hero row (and the tab bar/content below) when it loaded.
  Pinned the charm's 2812x2024 aspect ratio so the width is reserved up
  front.

Funnel split (was ~630 lines):
- givebackFunnelTypes.ts: STEP_KEYS / StepKey / CauseSelection
- GivebackFunnelVideo.tsx: the docked explainer video
- GivebackFunnelSteps.tsx: the per-step content
- GivebackFunnel.tsx: the orchestrator shell (state, nav, footer, portal)
