+++ packages/shared/src/features/giveback/components/GivebackMascot.tsx
@@ -35,13 +35,16 @@ export const GivebackMascot = ({
         className="bg-accent-cabbage-default/20 absolute inset-0 m-auto size-3/4 rounded-full blur-3xl motion-safe:animate-glow-pulse"
       />
       {/* The render sits on solid black; `mix-blend-screen` drops the black so
-          the charm reads as floating on the dark page. */}
+          the charm reads as floating on the dark page. The charm artwork is
+          2812x2024; pinning that aspect ratio reserves the width from the fixed
+          height before the image loads, so it never reflows the hero row (and
+          the tab bar/content below it) on load. */}
       <img
         src={image.src}
         alt={image.alt}
         loading="lazy"
         className={classNames(
-          'relative h-44 w-auto select-none object-contain mix-blend-screen motion-safe:animate-mascot-bob tablet:h-56',
+          'relative aspect-[703/506] h-44 w-auto select-none object-contain mix-blend-screen motion-safe:animate-mascot-bob tablet:h-56',
           imageClassName,
         )}
       />
